requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
idna==3.10
imbalanced-learn==0.13.0
imblearn==0.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
PubChemPy==1.0.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
rdkit==2025.3.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
sklearn-compat==0.1.3
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
